:truck: Move: signUpApi, profileApi 이동
Move: signUpApi, profileApi 이동
