--- app/_apis/signUp.ts
@@ -1,8 +1,0 @@
-import httpClient from "./httpClient";
-import { SignUpRequest, SignUpResponse } from "./schema/user";
-
-export const signUpApi = {
-  postSignUp: async (userData: SignUpRequest): Promise<SignUpResponse> => {
-    return await httpClient().post<SignUpResponse, SignUpRequest>("/signUp", userData);
-  },
-};
