+++ app/_components/Modal/SignUpModal/SignUpForm.tsx
@@ -2,7 +2,7 @@ import { SubmitHandler, useForm } from "react-hook-form";
 import { useMutation } from "@tanstack/react-query";
 import { SignUpRequest } from "@/app/_apis/schema/user";
 import useToggleHook from "@/app/_hooks/useToggleHook";
-import { signUpApi } from "@/app/_apis/signUp";
+import { signUpApi } from "@/app/_apis/userApi";
 import { useLogin } from "@/app/_context/LoginProvider";
 import { setToken } from "@/app/_utils/handleToken";
 import Input from "../../Input/Input";
